requirements.txt
fastapi==0.78.0
numpy
opencv_python
Pillow==9.2.0
tensorflow==2.8.0
uvicorn==0.16.0
